Bug fix to GEM interface for P3_v4. (4.0.17)
This includes modifications to s/r 'wrapepr_gem' (the proper construction of
the array 'diag_effi') plus corrections to the optional arguments in the
call statement to 'wrapper_gem' in 'condensation.F90'

LookupTable_2-v5.0 has problems -- tentatively, this has been reverted to v4.

Also, the interface files for GEM have been added to the repo.

TEST PERFORMED:

- This was tested in GEM (toy convective case) in multiple configruations
of 2/3-moment and nCat = 1,2,3