Updates to create_LT1.  Now, LT1-5.2-2momI and LT1-5.3-3momI.
Now, the version number of LT1 uniquely defines all parameters, except the switch
for 2momI or 3momI.  The last version of P3 (4.0.17) works with LT2-4 for nCat>1;
this version has Dm_max = 2000.  However, with 3-moment we use Dm_max=400000.[m]
in LT1 -- thus, there are different versions of LT1 for 2-mom and 3-mom.

Consequently, with 3momI, ice-ice interaction is suppressed. (LT2 is read but not used.)
A subsequent package will have a consistent Dm_max for all LTs.

TESTS:

- P3-4.0.17, LT1-5.2-2momI, LT1-5.3-3momI, and LT2-v4 have been tested in various
  configurations in GEM (toy model, 1 h run) for 2/3-mom and nCat=1,2,3.