Modifications to GEM wrapper. (v3.1.6.6)
This was a modification to the interface with the GEM physics in GEM/x/5.2.a??,
which now requires physics tendencies, rather than updated fields, to be passed back..
It was put in by Ron McTaggart-Cowan (RPN-A), prior to the migration to U2.
It was reported by Melissa Cholette that this was tested for bit-matching compared
to the previous version, at least for nCat=1.

It has not been tested for multiple ice categories (nCat>1) and it appears that
this was hard-coded for nCat=1 only.