Modified function 'maxHailSize'.
- made appropriate calculations in double-precision to avoid overflows
- switched to flux method as the default

Note: With the values of Ncrit and Rcrit in current version, these
methods give similar max. sizes.  The flux method is chosen since
its calibration is based on more physically observable (and easier
to visualize quantities).

Still need to optimize the use of this function (e.g. adding filters
based on density and mean size)