Set up v4 bug-fix branch, for continued archive of fixes off master.
The main purpose of this branch is allow for bug fixes post-v4.5.1
to be done on v4 which have been added since the merge to v5 on master.

There is no intention of merging these changes back to master. This
branch serves as a record of bug-fix versions to v4 which models
which have not yet switched to v5.