Added "generic" mp_p3_wrapper_cm1.
This was copied form mp_p3_wrapper_wrf_2cat. The objective is it
replace the two _wrf wrappers, also used for CM1, with a single
generlized (and cleaned-up) wrapper for all configurations of P3
for use in CM1, and for consideration in WRF.

In mp_driver.F (CM1), calls to mp_p3_wrapper_wrf (ptype 53,60)
and to mp_p3_wrapper_wrf_2cat (ptype 61,62,63) are now done
with calls to mp_p3_wrapper_cm1.

At this stage, the results bit-match in CM1 for tests done with
ptype 53, 60, 61, 62, 63.