Tidy-up (non-answer-changing) code for strict diam_ice calculation.
This code corresponds to v61 for CM1-hail study.

Tested in CM1 for all configs/cases.  Appears to systematically
increase Ze in core and slightly reduce ice sizes aloft.

For GEM, Melissa noted a crash (for Calgary hailstorm case), so
this will need to be investigated in GEM before mergin to MASTER.