Modification to diam_ice computation for multi-cat processes
We changed it to use the mean mass-weighted density to compute
the equivalent-spherical diameter instead of taken the mean
mass-weighted diameter from the lookup table because the latter
was giving to large ice crystals.
This has been tested in 1D and in GEM and will only impacts
solutions with nCat>1. Future work will be to improve the
whole merging/initiation of ice category with multi-cat.