Added interface for GEM/x/5.2.0-rc1.
The modified files are identical to those of 5.2.0-b1; however,
microphy_p3.F90 had to be modified to work with -rc1 to accomodate
to a GEM bug in which ECCCGEM is not defined.

Tested in GEM 5.2.0-rc1.  This version bit-matches runs done
with P3_v5.3.0, for various TM/LF/nCat combos.