Changed 'thres_raindrop' from 150 back to 100 microns. (now v5.3.2)
Melissa tested this and concluded that 100 microns was better to
distinguish rain from drizzle (for pcp type diagnostic only).

Also, "added" GEM interface for x/5.2.0-rc3, which contains the
bug fix for the env variable ECCCGEM (but with no changes to
the GEM physics -- thus, this is identical to rc1 but without
the optional "correction" to P3 to adapt to this GEM bug).