Bug fix to lookup table 1 (both 2momI and 3momI)
Fix to the melting equation when liquid fraction on only.
The melting term associated with the ventilation coefficient
was 1 instead of 0 for particles smaller than 100 um in the PSD.

The fix should slow the melting a little bit as shown with the 1D model.
The code has been tested in a full series of cases for winter 2022 (HRDPS).