Bugfix to the lambda limiter for the ice size distribution
Because of high values of mass-weighted diameter (Dmi) when Firim is low.
This bug was introduced in P3v4.
All lookup tables have been updated.
This code has been tested for series (winter and summer) of cases in the HRDPS.
The impacts are small (see pptx on the git issue).