Added modifications to GEM-specific functions from GEM_5.3.0-a10.
Modifications were made (by RPN-A) to P3_5.3.6 in GEM_5.3.0-a9,
appearing in -a10. These were to functions 'p3_lwc' and 'p3_iwc'
and only impact GEM (and are not even compiled outside of GEM).

Note, therefore, that the version of P3 in GEM_5.3.0-a10 is
different than P3_5.3.6 in the P3 repository, despite the unmodified
comment block (and P3 version number) at the top of the module.
This is a (hopefully final) deviation from the normal P3
development workflow.