Clean-up of cld1d.f90 and added timing functionality.
Major clean-up of old, commented, etc. code to cld1d.f90
  + changed variable names for tracers to P3 names
  + removed unused write statements
  + added compiler option to use old versions of P3.  With this,
    we can maintain a single version of cld1d.f90 but test old
    versions of P3 if necessary

Added functionality to time specific blocks of P3 code.  The timing
code is wrapped in compiler variable #ifdef statements, so this
code can stay in (it will not impact 3D runs, e.g.). This will allow
us to keep an eye on the computational cost as changes are made.

Modified Makefile for kin1D system
  + compiled in single precision
  + 'timing' variable is defined at compile time

--> bitmatches baseline P3 v5.3.14