Commented call to maxHailSize in type-diagnostics section.
This function is very expensive -- in cld1d, it adds 0.5% to the
total run time, even though it is only called every 5 min for
the full column (and every time step at the lowest level). In
the future, this field could be added to LUT_1.

Also, cleaned up WRF/CM1-specific diagnostic code block.