Replaced iterative algorithm to solve mu_i with optimized one.
This is done in the new subroutine 'solve_mui'.  The algorithm
does the same thing as before, but is coded better so as to
exit loop upon convergence.  As a result, fewer iterations
are typically done (often only 2, vs. the hard-wired 5 previously)

Tests were done with various initial densities and convergence
criteria.  rho_i = 200 kg/m3 and a tolerance of 0.25 for mu_i
between successive iterations seems to work well.

The results change, but the plots are practically identical.
The total cost of p3_main in kin1d is reduced by 15%.