Minor bug fix to 'compute_mu_3moment'
In the piecewise polynomial, for small values of G (~ 1.6) the
polynomial expression (second last one) gave large values of mu_i.

This changes the solution very slightly, but the plots are
indistinguishable.

Note -- this bug may have been corrected independently by Hugh
for dev-full3moment); upon merging (for v5.4) we will need to
make sure the best bug-fix is used.

Minor bug fix to 'compute_mu_3moment'

In the piecewise polynomial, for small values of G (~ 1.6) the
polynomial expression (second last one) gave large values of mu_i.

This changes the solution in kin1D very slightly, but the plots are
indistinguishable.

The total run time is ~1% faster.