Optmization of type diagnostics
Replaced code like:

 if (Dr < thres_raindrop) then
   Q_drizzle(i,k) = qr(i,k)
     else
   Q_rain(i,k)    = qr(i,k)
 endif

with

 tmp2 = merge(1., 0., Dr < thres_raindrop) !1. for drizzle, 0. for rain
 Q_drizzle(i,k) = qr(i,k)*tmp2
 Q_rain(i,k)    = qr(i,k)*(1.-tmp2)

The goal is to replace the conditionals with the more efficent code.

Bit-matches in kin1D; however I have not checked in detail by
printing Q_drizzle(i,k), etc.

Also:  freq3DtypeDiag was reduced from 5. (minutes) to 60 min.
This will reduce the costs of full 3D runs (long duration).  Probably
this, combined with maxHailSize, was causing a notable cost increase
for 3-moment in GEM (etc.) tests with v5.