Optimized 'access_lookup_table_3mom_LF'; now also used for 3mom_noLF
Uncommented existing code to skip interpolation of LUT over liq-frac
if qiliq=0 (with a slight modification to the condition.  This speeds
up the interpolation for the LF_on config, since this conditions is
often met, and also allows this subroutine to be used for 3mom-LF_off.

Consequently, this completely removes the need for the subroutine
'access_lookup_table_3mom'; now all calls in p3_main are done to the
same one, even if log_liqfrac = .false.

Tests in kin1D:
- 3mom, LF_off:  reduces total cost of p3_main by 20%
- 3mom, LF_on :  reduces total cost of p3_main by 15%
(results bit-match)

Note: The same optimization can and will be done for the 2mom_LF LUT.