Clean-up of cld1d.f90 and added timing functionality.
Major clean-up of old, commented, etc. code to cld1d.f90
  + changed variable names for tracers to P3 names
  + removed unused write statements
  + added compiler option to use old versions of P3.  With this,
    we can maintain a single version of cld1d.f90 but test old
    versions of P3 if necessary

Added functionality to time specific blocks of P3 code.  The timing
code is wrapped in compiler variable #ifdef statements, so this
code can stay in (it will not impact 3D runs, e.g.). This will allow
us to keep an eye on the computational cost as changes are made.

Modified Makefile for kin1D system
  + compiled in single precision
  + 'timing' variable is defined at compile time

Also, renamed (temporary) version to '5.4.3 + optimize', ultimately
to be closed as v5.4.4 on the 'development' branch.  Note, this is
in vioalation of the naming convention for the P3 workflow protocol;
this was due to the fact that dev-full3moment had beed closed at
one point (as v5.4.0) but then removed from the 'development' branch
in order to advance v5.3.x, prior to making v5.4.0 public.