Optimized 'access_lookup_table_3mom_LF'.
Uncommented existing optimization code for conditions with liqFrac=0.
This speeds up the code considerably for these conditions (which
are common with LF on) and also allows for this routine to be used,
without any extra cost, for 3-mom w/ LF off.  As such, all of the calls
to 'access_lookup_table_3mom' have been replaced by calls to
'access_lookup_table_3mom_LF', thereby removing the need for several
'IF (log_liqfrac) THEN ...' statements.

Results in kin1D bitmatch; the timing of 'p3_main' is reduced by
15-20%.

Note, the same can/will be done for other call-LUT subroutinesfor which
there are no-LF and LF versions