Replaced multiple code blocks to solve mu_i with a function.
Currently it just uses the same (sub-optimal) iteration to
solve for mu_i, but this sets the stage to optimize the function.

Bit-matches in kin1d.

Note - with the introduction of full-3-moment, there are still some
places where this new function could probably replace remaining
code blocks.

Also -- ice density (f1pr16) could be passed out of 'solve_mui'
to avoid calls to access_LUT.  (It was excluded since the goal is
to eventually replace 'solve_mui' with a dedicated LUT.)