Replaced iterative algorithm to solve mu_i with optimized one.
This is done in the new subroutine 'solve_mui'.  The algorithm
does the same thing as before, but is rearranged and coded to
exit loop upon convergence.  As a result, fewer iterations
are typically done (often only 2, vs. the hard-wired 5 previously)

Tests were done with various initial densities and convergence
criteria.  A tolerance of 0.25 for mu_i between successive
iterations seems to work well.

The results in kin1D change, but the plots are practically identical.
The total cost of p3_main in kin1d is reduced by around 5%