Code clean-up plus test summary document. Closing as: v5.3.15
New doc:  ./test-summaries/20250505-v5.3.15_optimization_kin1D.pdf

In kin1D, timing tests for 3mom-LF give a reduction in the total time
of 'p3_main' of 25%.  (This includes avoiding the costly maxHailSize
function, but other optimizations also.

Note: The optimization is not complete, but this will be continued
in a development branch off of v5.4.(3), once dev-full3mom is merged
to 'main'.