Updated version number to v5.4.3 in preperation for  merging to 'main'.
Note, there was some confusion about the version numbers for the
dev-full3mom branch since it had, at one point, been merged to 'main'
(formerly 'development') and subsequently advanced (to v5.4.1, etc.)
However, due to the need to make other changes to v5.3.x, we "unmerged"
the full3mom, keeping the dev-full3mom branch active, and made further
changes to v5.3.x.  Now, upon merging dev-full3mom to 'main', the
closed version will be v5.4.3. This follows from Melissa requiring to
publish (Zenodo) a copy of the code she used for her 2025 MWR paper
(which was technically was 'v5.4.1 + changed', which we opted to name
'v5.4.2'.