Corrected code for zitot updates for old group 1 processes.
Previously this code block was the same as for the new (full) approach.
Now, that section updates zitot based on the constant-mu_i (for
those processes only) closure assumptions.

The kin1D results for log_full3Mom = .false. are now very similar
to those from v5.3.15.  The computational cost is now about the same
(which is about 10-15% faster than with log_full3Mom = .true.