Added functionality to kin1d to automate test sets in kin1d.
Each commit with changes to P3 should include this full set of tests.
It is a one-line command with the script 'run_full_tests'. After
compiling kin1d, various configuraitons are run:

1cat-2momI-noLF
1cat-3mom-LF
2cat-3mom-LF
3cat-3mom-LF

Plots are then automatically generated.  The tests from the previous
commit are transfered to './results-previous_commit' and then all
output files, plots, timing files, and code are moved/copied to
./results'.

This will facilitate catching any changes to the results and/or
timing along the development process (rather than having to backtrack).