Bugfix to 2 conditional statements (bugfix from Vivian Lee, RPN-A)
There were two conditionals of the form:

  if (qitot(i,k,iice).ge.qsmall .and. qiliq(i,k,iice)/qitot(i,k,iice).lt.0.01) then

For newer Intel compilers, this can fail since the calculation with
a possible divide-by-zero is made with the IF statement.  The statements
have been modified to be of the form:

  if (qitot(i,k,iice).ge.qsmall) then
     if (qiliq(i,k,iice)/qitot(i,k,iice).lt.0.01) then

Note, his has never created any known crashes, but apparently for the
new Intel Fortran compiler at RPN this does not pass compilation.