Bugfix (undeclared variables).  New version: 5.4.5
In the previous version (the last commit), a minor bugfix was made
by cherry-picking the bugfix from a later development version.
However, the code in the later version also had some changes to
some variable names, so the cherry-pick introduced a new problem
of undeclared variables.

Tested in kin1d.

Evidently, the v5.4.4 was not tested before pushing it to the repository.
This is not standard proceedure -- normally we test all changes,
even if they are small. Note, in a future version test results for
every commit will be automated (in kin1d) and documented.