Modified s/r 'impose_maximum_total_Ni' to prevent reduction of already low values.
This address issue #61.

In the previous method, the total number (of all categories) was limited to 'max_total_Ni'
by scaling nitot of each category to limit the sum.  However, this could reduce the
number in a category with already a low value, thereby creating large mean sizes and
correspondingly large reflectivity values.

In the modification, each category is individually limited to 'max_total_Ni'.
This means that it is technically possible for the total to equal nCat*max_total_Ni.
However, this is unlikely, and furthermore this maximum parameter value is arbitrary.

The s/r name has been changed to 'impose_maximum_Ni' and the paramter 'max_total_Ni'
has been changed to 'max_Ni' (i.e. per ice category).

Tested in kin1d.  Zero change for nCat=1; tiny changes for nCat=2 or 3 (but the plots
are indistinguishable.

It is expected that this modification will solve the high reflectivity values that
Dan Dawson was getting in spring 2025; however, this has not yet been tested in 3D.