Bugfix -- NaNs in precip rate due to uninitialzed value of sfc mass flux.
This bug was introduced as a result of an optimization to replace a
conditional [IF (k.eq.k_bot) ...] with a MERGE statement in the new
sedimentation subroutines. The new code, recently implemented, resulted
in NaNs due to updating surface precipitation accumulation for situations
where the mass flux at the bottom level was not computed.  The fix was
to initialize the mass flux at this level to 0. at the start of the
subroutine.

Tested in CM1 for all ptypes (50-54).