Minor cleanup and adding limit of zlarge for zitot:  v5.5.0-rc9
The line

  zitot = max(zitot,zsmall)

that preceded each (but one) call to subroutine 'get_mui_rhoi' was
modified to be

  zitot = max(zsmall, min(zlarge, zitot))

and was relocated to be inside of 'get_mui_rhoi' (which now returns
the modified value of zitot).  The relocation was done simply to
simplify the code.

Limiting to a maximum of zlarge (1.) is done in response to a
crash in the HRDPS testing with 5.5.0-rc8 which resulted from an
overflow in the calculation dum8 = (alog10(zitot/qitot)+23.)*3.10347652
in 'find_lookupTable_indices_3a' where the value of zitot was very large
(> 1.e24) and qitot was small (~ 2.e-14).  The reason for the large
zitot is likely due to advection of small values (recall: the advected
dynamic variable is not actually zitot but gets converted to zitot
via 'zitot = zitot**2/nitot' at the beginning of p3_main, before any
ajustment to nitot.

Also, the default value of 'log_full3momi' was changed to .false.
The results are very similar and the integration is slightly faster.

- Tested in CM1 (quarter_ss); plots bitmatch for comparisons of rc9
  and rc8 (modified with log_full3momi = .false).

- Tested in GEM for case that crashed -- works now.